--- a/orderbook/src/uint64_hashmap.c
+++ b/orderbook/src/uint64_hashmap.c
@@ -3,34 +3,43 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <time.h>
 
-int _uint64_hashmap_probe(struct uint64_hashmap* map, uint64_t key);
+uint64_t uint64_hash(uint64_t key) {
+  key = (key ^ (key >> 30)) * UINT64_C(0xbf58476d1ce4e5b9);
+  key = (key ^ (key >> 27)) * UINT64_C(0x94d049bb133111eb);
+  key = key ^ (key >> 31);
+  return key;
+}
+
+uint32_t find_next_positive_power_of_two(uint32_t value) {
+  uint8_t num_leading_zero_bits = 0;
+  value--;
+
+  while ((value & (1 << (UINT32_WIDTH - 1))) == 0) {
+    num_leading_zero_bits++;
+    value <<= 1;
+  }
+
+  return 1 << (UINT32_WIDTH - num_leading_zero_bits);
+}
+
+void _uint64_hashmap_compact_chain(struct uint64_hashmap* map,
+                                   uint64_t delete_index);
 void _uint64_hashmap_resize(struct uint64_hashmap* map, uint32_t capacity);
 
 struct uint64_hashmap uint64_hashmap_with_capacity(uint32_t capacity) {
-  srand(time(NULL));
-  uint32_t prime = DEFAULT_PRIME_FACTOR;
+  capacity = find_next_positive_power_of_two(capacity);
 
   // initialise the table to be empty
   struct uint64_hashmap_entry* table =
-      malloc(sizeof(struct uint64_hashmap_entry) * capacity);
-  for (int i = 0; i < capacity; i++)
-    table[i].key = DEFAULT_EMPTY_KEY;
+      calloc(capacity, sizeof(struct uint64_hashmap_entry));
 
-  struct uint64_hashmap map = {.max_load_factor = DEFAULT_MAX_LOAD_FACTOR,
-                               .capacity = capacity,
-                               .prime = prime,
-                               // required to make the bound one less and +1
-                               // to make sure scale % prime != 0
-                               .scale = (rand() % (prime - 1)) + 1,
-                               .shift = rand() % prime,
-                               .table = table};
+  struct uint64_hashmap map = {.capacity = capacity, .table = table};
   return map;
 }
 
 struct uint64_hashmap uint64_hashmap_new() {
-  return uint64_hashmap_with_capacity(DEFAULT_CAPACITY);
+  return uint64_hashmap_with_capacity(UINT64_HASHMAP_DEFAULT_CAPACITY);
 }
 
 void uint64_hashmap_free(struct uint64_hashmap* map) {
@@ -40,48 +49,63 @@ void uint64_hashmap_free(struct uint64_hashmap* map) {
 void* uint64_hashmap_put(struct uint64_hashmap* map,
                          uint64_t key,
                          void* value) {
-  if (key == TOMBSTONE_KEY) {
-    fprintf(stderr, "key cannot be TOMBSTONE_KEY(%ld)\n", TOMBSTONE_KEY);
-    exit(EXIT_FAILURE);
+  if (value == NULL)
+    return NULL;
+
+  const uint64_t mask = map->capacity - 1;
+  uint64_t index = uint64_hash(key) & mask;
+
+  void* previous_value;
+
+  while ((previous_value = map->table[index].value) != NULL) {
+    if (key == map->table[index].key)
+      break;
+    index = (index + 1) & mask;
   }
 
-  int i = _uint64_hashmap_probe(map, key);
-  if (i >= 0) {  // found existing, update
-    void* previous_value = map->table[i].value;
-    map->table[i] = (struct uint64_hashmap_entry){.key = key, .value = value};
-    return previous_value;
+  if (previous_value == NULL) {
+    ++map->size;
+    map->table[index].key = key;
   }
-  map->table[-(i + 1)] =
-      (struct uint64_hashmap_entry){.key = key, .value = value};
-  map->size++;  // increment the current size
 
-  // Since linear probing works best when the load factor is low, we
-  // _uint64_hashmap_resize the table once the load factor exceeds this
-  // specified threshold.
-  uint32_t load_factor =
-      ((map->size + map->tombstone_count) * 100) / map->capacity;
-  if (load_factor > map->max_load_factor)
-    _uint64_hashmap_resize(map, 2 * map->capacity - 1);
+  map->table[index].value = value;
 
-  return NULL;
+  if (map->size > (UINT64_HASHMAP_MAX_LOAD_FACTOR * map->capacity) / 100)
+    _uint64_hashmap_resize(map, map->capacity << 1);
+
+  return previous_value;
 }
 
 void* uint64_hashmap_get(struct uint64_hashmap* map, uint64_t key) {
-  int i = _uint64_hashmap_probe(map, key);
-  if (i < 0)  // unable to find the key
-    return NULL;
-  return map->table[i].value;
+  const uint64_t mask = map->capacity - 1;
+  uint64_t index = uint64_hash(key) & mask;
+
+  void* value;
+  while ((value = map->table[index].value) != NULL) {
+    if (key == map->table[index].key)
+      break;
+    index = (index + 1) & mask;
+  }
+
+  return value;
 }
 
 void* uint64_hashmap_remove(struct uint64_hashmap* map, uint64_t key) {
-  int i = _uint64_hashmap_probe(map, key);
-  if (i < 0)  // not able to find the element to remove
-    return NULL;
-  void* removed = map->table[i].value;
-  map->table[i] = (struct uint64_hashmap_entry){.key = TOMBSTONE_KEY};
-  map->size--;
-  map->tombstone_count++;
-  return removed;
+  const uint64_t mask = map->capacity - 1;
+  uint64_t index = uint64_hash(key) & mask;
+
+  void* value;
+  while ((value = map->table[index].value) != NULL) {
+    if (key == map->table[index].key) {
+      map->table[index].value = NULL;
+      map->size--;
+      _uint64_hashmap_compact_chain(map, index);
+      break;
+    }
+    index = (index + 1) & mask;
+  }
+
+  return value;
 }
 
 char* uint64_hashmap_print(struct uint64_hashmap* map) {
@@ -90,8 +114,7 @@ char* uint64_hashmap_print(struct uint64_hashmap* map) {
 
   len += sprintf(str + len, "{");
   for (int i = 0; i < map->capacity; i++) {
-    if (map->table[i].key == DEFAULT_EMPTY_KEY ||
-        map->table[i].key == TOMBSTONE_KEY)
+    if (map->table[i].value == NULL)
       continue;
 
     len += sprintf(str + len, "%ld[%d]: %p, ", map->table[i].key, i,
@@ -106,14 +129,29 @@ char* uint64_hashmap_print(struct uint64_hashmap* map) {
   return str;
 }
 
-/**
- * Search the table in a cyclic fashion to find if a key exist. The function
- * returns positive integer if a matching key is found, the returned value is
- * the index. Otherwise, it returns a negative integer indicating a free slot in
- * the table.
- */
-int _uint64_hashmap_probe(struct uint64_hashmap* map, uint64_t key) {
-  MAP_PROBE(map, key)
+void _uint64_hashmap_compact_chain(struct uint64_hashmap* map,
+                                   uint64_t delete_index) {
+  const uint64_t mask = map->capacity - 1;
+  uint64_t index = delete_index;
+
+  while (true) {
+    index = (index + 1) & mask;
+    void* value = map->table[index].value;
+    if (value == NULL)
+      break;
+
+    const uint64_t key = map->table[index].key;
+    const uint64_t hash = uint64_hash(key) & mask;
+
+    if ((index < hash && (hash <= delete_index || delete_index <= index)) ||
+        (hash <= delete_index && delete_index <= index)) {
+      map->table[delete_index].key = key;
+      map->table[delete_index].value = value;
+
+      map->table[index].value = NULL;
+      delete_index = index;
+    }
+  }
 }
 
 void _uint64_hashmap_resize(struct uint64_hashmap* map, uint32_t capacity) {
@@ -128,15 +166,11 @@ void _uint64_hashmap_resize(struct uint64_hashmap* map, uint32_t capacity) {
   map->capacity = capacity;
 
   // Copy back the items in temp into the new
-  map->table = malloc(sizeof(struct uint64_hashmap_entry) * capacity);
-  for (int i = 0; i < capacity; i++)
-    map->table[i].key = DEFAULT_EMPTY_KEY;
+  map->table = calloc(capacity, sizeof(struct uint64_hashmap_entry));
   map->size = 0;
   for (int i = 0; i < old_capacity; i++)
-    if (temp[i].key != DEFAULT_EMPTY_KEY && temp[i].key != TOMBSTONE_KEY)
+    if (temp[i].value != NULL)
       uint64_hashmap_put(map, temp[i].key, temp[i].value);
-
-  map->tombstone_count = 0;  // reset tombstone count
 
   free(temp);
 }