On the `main` branch which uses a naive implementation of hashmaps in the `uint64_hashmap.c`, the benchmark was:

```sh
Over 100 samples,
Took 56.89ms to process 100000 messages
Took 568ns to process 1 message
-------------------------------
orderbook_market: 102ns/op over 44 calls
orderbook_limit: 746ns/op over 49706 calls
orderbook_cancel: 397ns/op over 49777 calls
orderbook_amend_size: 54ns/op over 473 calls
```

In this PR, I referenced how [Agrona](), a high-performance data structures library in Java implemented its [Int2ObjectHashMap](https://github.com/real-logic/agrona/blob/master/agrona/src/main/java/org/agrona/collections/Int2ObjectHashMap.java), a specialized hashmap for 64bit integers as key and reference as values and rewrote `uint64_hashmap`, it achieves a result of:

```sh
Over 100 samples,
Took 14.82ms to process 100000 messages
Took 148ns to process 1 message
-------------------------------
orderbook_market: 86ns/op over 44 calls
orderbook_limit: 138ns/op over 49706 calls
orderbook_cancel: 159ns/op over 49777 calls
orderbook_amend_size: 43ns/op over 473 calls
```

Also, it achieves this is lesser LOC and thus I'm able to shave off some source code.